refactored code; added ribbon
