+++ lib/constants/strings/home_page_strings.dart
@@ -2,3 +2,4 @@ const String callUsText = "Call Us:";
 const String phoneNumberText = "+91-97118 24118";
 const String emailUsText = "Email Us";
 const String emailText = "contactus@companyraahi.com";
+const String payNowText = 'PAY NOW';
