+++ lib/constants/theme/text_styles.dart
@@ -4,7 +4,5 @@ const TextStyle callUsTextStyle = TextStyle(
   fontSize: 12,
   fontWeight: FontWeight.bold,
 );
-
-const TextStyle phoneNumTextStyle = TextStyle(
-  fontSize: 12,
-);
+const TextStyle phoneNumTextStyle = TextStyle(fontSize: 12);
+const TextStyle payNowTextStyle = TextStyle(fontSize: 16);
