+++ lib/pages/home_page/widgets/header_ribbon.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:sample/constants/strings/home_page_strings.dart';
-import 'package:sample/shared/widgets/ribbon_widget.dart';
+import 'package:sample/pages/home_page/widgets/ribbon_widget.dart';
+import 'package:sample/shared/widgets/custom_rounded_button.dart';
+
+import 'button_row.dart';
 
 class HeaderRibbon extends StatelessWidget {
   const HeaderRibbon({
@@ -26,55 +28,8 @@ class HeaderRibbon extends StatelessWidget {
           ),
           Row(
             children: [
-              TextButton(
-                style: ButtonStyle(
-                  padding: MaterialStateProperty.all(
-                    EdgeInsets.symmetric(vertical: 5.0, horizontal: 12.0),
-                  ),
-                  foregroundColor:
-                      MaterialStateProperty.all<Color>(Colors.white),
-                  backgroundColor:
-                      MaterialStateProperty.all<Color>(Color(0xff43C6AC)),
-                  shape: MaterialStateProperty.all(
-                    RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(50),
-                    ),
-                  ),
-                ),
-                onPressed: () {},
-                child: Padding(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 12.0, vertical: 4),
-                  child: Text(
-                    'PAY NOW',
-                    style: TextStyle(fontSize: 16),
-                  ),
-                ),
-              ),
-              IconButton(
-                icon: FaIcon(FontAwesomeIcons.facebook),
-                onPressed: () {},
-                color: Colors.blue[900],
-                iconSize: 30,
-              ),
-              IconButton(
-                icon: FaIcon(FontAwesomeIcons.linkedin),
-                onPressed: () {},
-                color: Colors.blue[700],
-                iconSize: 30,
-              ),
-              IconButton(
-                icon: FaIcon(FontAwesomeIcons.whatsapp),
-                onPressed: () {},
-                color: Colors.green[700],
-                iconSize: 30,
-              ),
-              IconButton(
-                icon: FaIcon(FontAwesomeIcons.facebookMessenger),
-                onPressed: () {},
-                color: Colors.blue,
-                iconSize: 30,
-              ),
+              CustomRoundedButton(),
+              ButtonRow(),
             ],
           ),
         ],
