+++ lib/shared/widgets/custom_rounded_button.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:sample/constants/strings/home_page_strings.dart';
+import 'package:sample/constants/theme/text_styles.dart';
+
+class CustomRoundedButton extends StatelessWidget {
+  const CustomRoundedButton({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return TextButton(
+      style: ButtonStyle(
+        padding: MaterialStateProperty.all(
+          EdgeInsets.symmetric(vertical: 5.0, horizontal: 12.0),
+        ),
+        foregroundColor: MaterialStateProperty.all<Color>(Colors.white),
+        backgroundColor: MaterialStateProperty.all<Color>(Color(0xff43C6AC)),
+        shape: MaterialStateProperty.all(
+          RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(50),
+          ),
+        ),
+      ),
+      onPressed: () {},
+      child: Padding(
+        padding: const EdgeInsets.symmetric(
+          horizontal: 12.0,
+          vertical: 4,
+        ),
+        child: Text(payNowText, style: payNowTextStyle),
+      ),
+    );
+  }
+}
